+++ python/cppyy/__init__.py
@@ -366,7 +366,10 @@ def sizeof(tt):
     try:
         return _sizes[tt]
     except KeyError:
-        sz = gbl.gInterpreter.ProcessLine("sizeof(%s);" % (_get_name(tt),))
+        if type(tt).__module__ == '_ctypes':
+            sz = ctypes.sizeof(tt)
+        else:
+            sz = gbl.gInterpreter.ProcessLine("sizeof(%s);" % (_get_name(tt),))
         _sizes[tt] = sz
         return sz